requirements.txt
# Core dependencies for Frankenstino AI
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
fastapi>=0.100.0
uvicorn>=0.23.0
numpy>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# LLM inference (GGUF models)
llama-cpp-python>=0.2.0

# Document processing
PyPDF2>=3.0.0
PyMuPDF>=1.22.0  # Better PDF processing
python-docx>=0.8.11
pandas>=1.5.0
openpyxl>=3.0.10
xlrd>=2.0.1
python-pptx>=0.6.21
ebooklib>=0.18
lxml>=4.9.0  # For XML/HTML processing
PyYAML>=6.0  # For YAML processing

# Optional GPU acceleration (uncomment if needed)
# faiss-gpu>=1.7.0

# Monitoring dependencies
prometheus-client>=0.17.0
psutil>=5.9.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
